agents/notification_agent/requirements.txt
redis>=4.5.1
openai>=0.27.0
requests>=2.28.0
python-dotenv>=1.0.0
crewai>=0.28.0
langchain-openai>=0.0.2
# Notification-specific dependencies
slack-sdk>=3.26.0
pagerduty-sdk>=0.3.0
webexteamssdk>=1.6.0
nats-py==2.4.0